# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "HyperHive",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "hyperhive.vajid.tech",
    "basePath": "/",
    "paths": {
...
        "/friend/block": {
            "get": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Retrieve blocked friend requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get Blocked Friend Requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Offset for paginated results",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of blocked friend requests",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
        "/friend/friends": {
            "get": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Retrieve user's friends.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get Friends",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Offset for paginated results",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's friends",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
        "/friend/received": {
            "get": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Retrieve received friend requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get Received Friend Requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Offset for paginated results",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of received friend requests",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
        "/friend/request": {
            "post": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Send a friend request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Send Friend Request",
                "parameters": [
                    {
                        "description": "Request body for sending friend request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel_friend_svc.FriendRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Friend request sent successfully",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
        "/friend/restrict": {
            "patch": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Update friendship status (e.g., restrict or unblock).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Update Friendship Status",
                "parameters": [
                    {
                        "description": "Request body for updating friendship status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel_friend_svc.FrendShipStatusUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friendship status updated successfully",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
        "/friend/send": {
            "get": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Retrieve sent friend requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get Sent Friend Requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Offset for paginated results",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit number of results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sent friend requests",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
        "/friendMessage/message": {
            "get": {
                "security": [
                    {
                        "UserAuthorization": []
                    }
                ],
                "description": "Retrieve chat messages.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get Chat Messages",
                "parameters": [
                    {
                        "description": "Request body for getting chat messages",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodel_friend_svc.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chat messages",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/responsemodel_friend_svc.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestmodel_friend_svc.ChatRequest": {
            "type": "object",
            "required": [
                "friendID",
                "limit",
                "offset"
            ],
            "properties": {
                "friendID": {
                    "type": "string"
                },
                "limit": {
                    "type": "string"
                },
                "offset": {
                    "type": "string"
                }
            }
        },
        "requestmodel_friend_svc.FrendShipStatusUpdate": {
            "type": "object",
            "properties": {
                "FrendShipID": {
                    "type": "string"
                }
            }
        },
        "requestmodel_friend_svc.FriendRequest": {
            "type": "object",
            "required": [
                "FriendID"
            ],
            "properties": {
                "FriendID": {
                    "type": "string"
                }
            }
        },
...
        "responsemodel_friend_svc.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "result": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAutherisation": {
            "type": "apiKey",
            "name": "AccessToken",
            "in": "header"
        },
        "UserAuthorization": {
            "type": "apiKey",
            "name": "AccessToken",
            "in": "header"
        },
        "UserConfirmToken": {
            "type": "apiKey",
            "name": "ConfirmToken",
            "in": "header"
        }
    }
}